Dev docs

The docstrings on this page describe internals, they are not part of the public API.

Graph storage

SparseMatrixColorings.Graph — Type

Graph{T}

Undirected graph structure stored in Compressed Sparse Column (CSC) format.

Fields

colptr::Vector{T}: same as for SparseMatrixCSC

rowval::Vector{T}: same as for SparseMatrixCSC

SparseMatrixColorings.BipartiteGraph — Type

BipartiteGraph{T}

Undirected bipartite graph structure stored in bidirectional Compressed Sparse Column format (redundancy allows for faster access).

A bipartite graph has two "sides", which we number 1 and 2.

Fields

g1::Graph{T}: contains the neighbors for vertices on side 1

g2::Graph{T}: contains the neighbors for vertices on side 2

SparseMatrixColorings.vertices — Function

vertices(bg::BipartiteGraph, Val(side))

Return the list of vertices of bg from the specified side as a range 1:n.

SparseMatrixColorings.neighbors — Function

neighbors(bg::BipartiteGraph, Val(side), v::Integer)

Return the neighbors of v (a vertex from the specified side, 1 or 2), in the graph bg.

SparseMatrixColorings.adjacency_graph — Function

adjacency_graph(H::AbstractMatrix)

Return a Graph representing the nonzeros of a symmetric matrix (typically a Hessian matrix).

The adjacency graph of a symmetrix matric A ∈ ℝ^{n × n} is G(A) = (V, E) where

V = 1:n is the set of rows or columns i/j

(i, j) ∈ E whenever A[i, j] ≠ 0 and i ≠ j

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

SparseMatrixColorings.bipartite_graph — Function

bipartite_graph(J::AbstractMatrix)

Return a BipartiteGraph representing the nonzeros of a non-symmetric matrix (typically a Jacobian matrix).

The bipartite graph of a matrix A ∈ ℝ^{m × n} is Gb(A) = (V₁, V₂, E) where

V₁ = 1:m is the set of rows i

V₂ = 1:n is the set of columns j

(i, j) ∈ E whenever A[i, j] ≠ 0

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

Low-level coloring

SparseMatrixColorings.partial_distance2_coloring — Function

partial_distance2_coloring(bg::BipartiteGraph, ::Val{side}, order::AbstractOrder)

Compute a distance-2 coloring of the given side (1 or 2) in the bipartite graph bg and return a vector of integer colors.

A distance-2 coloring is such that two vertices have different colors if they are at distance at most 2.

The vertices are colored in a greedy fashion, following the order supplied.

See also

BipartiteGraph

AbstractOrder

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005), Algorithm 3.2

SparseMatrixColorings.symmetric_coefficient — Function

symmetric_coefficient(
    i::Integer, j::Integer,
    color::AbstractVector{<:Integer},
    group::AbstractVector{<:AbstractVector{<:Integer}},
    S::AbstractMatrix{Bool}
)

symmetric_coefficient(
    i::Integer, j::Integer,
    color::AbstractVector{<:Integer},
    star_set::StarSet
)

Return the indices (k, c) such that A[i, j] = B[k, c], where A is the uncompressed symmetric matrix and B is the column-compressed matrix.

The first version corresponds to algorithm DirectRecover1 in the paper, the second to DirectRecover2.

References

Efficient Computation of Sparse Hessians Using Coloring and Automatic Differentiation, Gebremedhin et al. (2009), Figures 2 and 3

SparseMatrixColorings.star_coloring — Function

star_coloring(g::Graph, order::AbstractOrder)

Compute a star coloring of all vertices in the adjacency graph g and return a tuple (color, star_set), where

color is the vector of integer colors

star_set is a StarSet encoding the set of 2-colored stars

A star coloring is a distance-1 coloring such that every path on 4 vertices uses at least 3 colors.

The vertices are colored in a greedy fashion, following the order supplied.

See also

Graph

AbstractOrder

References

New Acyclic and Star Coloring Algorithms with Application to Computing Hessians, Gebremedhin et al. (2007), Algorithm 4.1

SparseMatrixColorings.acyclic_coloring — Function

acyclic_coloring(g::Graph, order::AbstractOrder)

Compute an acyclic coloring of all vertices in the adjacency graph g and return a vector of integer colors.

An acyclic coloring is a distance-1 coloring with the further restriction that every cycle uses at least 3 colors.

The vertices are colored in a greedy fashion, following the order supplied.

See also

Graph

AbstractOrder

References

New Acyclic and Star Coloring Algorithms with Application to Computing Hessians, Gebremedhin et al. (2007), Algorithm 3.1

SparseMatrixColorings.group_by_color — Function

group_by_color(color::Vector{Int})

Create group::Vector{Vector{Int}} such that i ∈ group[c] iff color[i] == c.

Assumes the colors are contiguously numbered from 1 to some cmax.

SparseMatrixColorings.get_matrix — Function

get_matrix(result::AbstractColoringResult)

Return the matrix that was colored.

SparseMatrixColorings.StarSet — Type

StarSet

Encode a set of 2-colored stars resulting from the star_coloring algorithm.

Fields

star::Dict{Tuple{Int64, Int64}, Int64}: a mapping from edges (pair of vertices) to their star index

hub::Vector{Int64}: a mapping from star indices to their hub (the hub is 0 if the star only contains one edge)

SparseMatrixColorings.TreeSet — Type

TreeSet

Encode a set of 2-colored trees resulting from the acyclic_coloring algorithm.

Fields

forest::DataStructures.DisjointSets{Tuple{Int64, Int64}}: a forest of two-colored trees

Concrete coloring results

SparseMatrixColorings.DefaultColoringResult — Type

struct DefaultColoringResult{structure, partition, decompression, M} <: AbstractColoringResult{structure, partition, decompression, M}

Default storage for the result of a coloring algorithm, containing minimal information.

Fields

matrix::Any: matrix that was colored

color::Vector{Int64}: one integer color for each column or row (depending on partition)

group::Vector{Vector{Int64}}: color groups for columns or rows (depending on partition)

See also

AbstractColoringResult

SparseMatrixColorings.DirectSparseColoringResult — Type

struct DirectSparseColoringResult{structure, partition, M} <: AbstractColoringResult{structure, partition, :direct, M}

Storage for the result of a coloring algorithm when the decompression target is a SparseMatrixCSC.

Fields

matrix::Any: matrix that was colored

color::Vector{Int64}: one integer color for each column or row (depending on partition)

group::Vector{Vector{Int64}}: color groups for columns or rows (depending on partition)

compressed_indices::Vector{Int64}: flattened indices mapping the compressed matrix B to the uncompressed matrix A: they satisfy nonzeros(A)[k] = vec(B)[compressed_indices[k]]

See also

AbstractColoringResult

Testing

SparseMatrixColorings.same_sparsity_pattern — Function

same_sparsity_pattern(A::AbstractMatrix, B::AbstractMatrix)

Perform a partial equality check on the sparsity patterns of A and B:

if the return is true, they might have the same sparsity pattern but we're not sure

if the return is false, they definitely don't have the same sparsity pattern

SparseMatrixColorings.directly_recoverable_columns — Function

directly_recoverable_columns(
    A::AbstractMatrix, color::AbstractVector{<:Integer}
    verbose=false
)

Return true if coloring the columns of the symmetric matrix A with the vector color results in a column-compressed representation that preserves every unique value, thus making direct recovery possible.

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

SparseMatrixColorings.symmetrically_orthogonal_columns — Function

symmetrically_orthogonal_columns(
    A::AbstractMatrix, color::AbstractVector{<:Integer};
    verbose=false
)

Return true if coloring the columns of the symmetric matrix A with the vector color results in a partition that is symmetrically orthogonal, and false otherwise.

A partition of the columns of a symmetrix matrix A is symmetrically orthogonal if, for every nonzero element A[i, j], either of the following statements holds:

the group containing the column A[:, j] has no other column with a nonzero in row i

the group containing the column A[:, i] has no other column with a nonzero in row j

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

SparseMatrixColorings.structurally_orthogonal_columns — Function

structurally_orthogonal_columns(
    A::AbstractMatrix, color::AbstractVector{<:Integer}
    verbose=false
)

Return true if coloring the columns of the matrix A with the vector color results in a partition that is structurally orthogonal, and false otherwise.

A partition of the columns of a matrix A is structurally orthogonal if, for every nonzero element A[i, j], the group containing column A[:, j] has no other column with a nonzero in row i.

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

Matrix handling

SparseMatrixColorings.respectful_similar — Function

respectful_similar(A::AbstractMatrix)
respectful_similar(A::AbstractMatrix, ::Type{T})

Like Base.similar but returns a transpose or adjoint when A is a transpose or adjoint.

SparseMatrixColorings.matrix_versions — Function

matrix_versions(A::AbstractMatrix)

Return various versions of the same matrix:

dense and sparse

transpose and adjoint

Used for internal testing.

Examples

SparseMatrixColorings.Example — Type

struct Example{TA<:(AbstractMatrix), TB<:(AbstractMatrix)}

Example coloring problem from one of our reference articles.

Used for internal testing.

Fields

A::AbstractMatrix: decompressed matrix

B::AbstractMatrix: column-compressed matrix

color::Vector{Int64}: vector of colors

SparseMatrixColorings.what_fig_41 — Function

what_fig_41()

Construct an Example from Figure 4.1 of "What color is your Jacobian?", where the nonzero entries are filled with unique values.

SparseMatrixColorings.what_fig_61 — Function

what_fig_61()

Construct an Example from Figure 6.1 of "What color is your Jacobian?", where the nonzero entries are filled with unique values.

SparseMatrixColorings.efficient_fig_1 — Function

efficient_fig_1()

Construct an Example from Figure 1 of "Efficient computation of sparse hessians using coloring and AD", where the nonzero entries are filled with unique values.

SparseMatrixColorings.efficient_fig_4 — Function

efficient_fig_4()

Construct an Example from Figure 4 of "Efficient computation of sparse hessians using coloring and AD", where the nonzero entries are filled with unique values.