API reference

Public, exported

SparseMatrixColorings.SparseMatrixColorings — Module

SparseMatrixColorings

SparseMatrixColorings.jl

Coloring algorithms for sparse Jacobian and Hessian matrices.

Getting started

To install this package, run the following in a Julia Pkg REPL:

pkg> add https://github.com/gdalle/SparseMatrixColorings.jl

Background

The algorithms implemented in this package are mainly taken from the following articles:

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

ColPack: Software for graph coloring and related problems in scientific computing, Gebremedhin et al. (2013)

Some parts of the articles (like definitions) are thus copied verbatim in the documentation.

Alternatives

ColPack.jl: a Julia interface to the C++ library ColPack

SparseDiffTools.jl: contains Julia implementations of some coloring algorithms

SparseMatrixColorings.GreedyColoringAlgorithm — Type

GreedyColoringAlgorithm <: ADTypes.AbstractColoringAlgorithm

Greedy coloring algorithm for sparse Jacobians and Hessians, with configurable vertex order.

Compatible with the ADTypes.jl coloring framework.

Constructor

GreedyColoringAlgorithm(order::AbstractOrder=NaturalOrder())

Implements

ADTypes.column_coloring and ADTypes.row_coloring with a partial distance-2 coloring of the bipartite graph

ADTypes.symmetric_coloring with a star coloring of the adjacency graph

Example use

using ADTypes, SparseMatrixColorings, SparseArrays

algo = GreedyColoringAlgorithm(SparseMatrixColorings.LargestFirst())
A = sparse([
    0 0 1 1 0
    1 0 0 0 1
    0 1 1 0 0
    0 1 1 0 1
])
ADTypes.column_coloring(A, algo)

# output

5-element Vector{Int64}:
 1
 2
 1
 2
 3

See also

AbstractOrder

Public, not exported

Orders

SparseMatrixColorings.AbstractOrder — Type

AbstractOrder

Abstract supertype for vertex ordering schemes.

Subtypes

NaturalOrder

RandomOrder

LargestFirst

SparseMatrixColorings.NaturalOrder — Type

NaturalOrder()

Order vertices as they come in the graph.

SparseMatrixColorings.RandomOrder — Type

RandomOrder(rng=default_rng())

Order vertices with a random permutation.

SparseMatrixColorings.LargestFirst — Type

LargestFirst()

Order vertices by decreasing degree.

Decompression

SparseMatrixColorings.decompress_columns! — Function

decompress_columns!(
    A::AbstractMatrix{R},
    S::AbstractMatrix{Bool},
    C::AbstractMatrix{R},
    colors::AbstractVector{<:Integer}
) where {R<:Real}

Decompress the thin matrix C into a fat matrix A with the same sparsity pattern as S.

Here, colors is a column coloring of S, while C is a compressed representation of matrix A obtained by summing the columns that share the same color.

SparseMatrixColorings.decompress_columns — Function

decompress_columns(
    S::AbstractMatrix{Bool},
    C::AbstractMatrix{R},
    colors::AbstractVector{<:Integer}
) where {R<:Real}

Decompress the thin matrix C into a new fat matrix A with the same sparsity pattern as S.

Here, colors is a column coloring of S, while C is a compressed representation of matrix A obtained by summing the columns that share the same color.

SparseMatrixColorings.decompress_rows! — Function

decompress_rows!(
    A::AbstractMatrix{R},
    S::AbstractMatrix{Bool},
    C::AbstractMatrix{R},
    colors::AbstractVector{<:Integer}
) where {R<:Real}

Decompress the small matrix C into the tall matrix A with the same sparsity pattern as S.

Here, colors is a row coloring of S, while C is a compressed representation of matrix A obtained by summing the columns that share the same color.

SparseMatrixColorings.decompress_rows — Function

decompress_rows(
    S::AbstractMatrix{Bool},
    C::AbstractMatrix{R},
    colors::AbstractVector{<:Integer}
) where {R<:Real}

Decompress the small matrix C into a new tall matrix A with the same sparsity pattern as S.

Here, colors is a row coloring of S, while C is a compressed representation of matrix A obtained by summing the columns that share the same color.

SparseMatrixColorings.color_groups — Function

color_groups(colors)

Return groups::Vector{Vector{Int}} such that i ∈ groups[c] iff colors[i] == c.

Assumes the colors are contiguously numbered from 1 to some cmax.

Private

Graphs

SparseMatrixColorings.Graph — Type

Graph{T}

Undirected graph structure stored in Compressed Sparse Column (CSC) format.

Fields

colptr::Vector{T}: same as for SparseMatrixCSC

rowval::Vector{T}: same as for SparseMatrixCSC

SparseMatrixColorings.BipartiteGraph — Type

BipartiteGraph{T}

Undirected bipartite graph structure stored in bidirectional Compressed Sparse Column format (redundancy allows for faster access).

A bipartite graph has two "sides", which we number 1 and 2.

Fields

g1::Graph{T}: contains the neighbors for vertices on side 1

g2::Graph{T}: contains the neighbors for vertices on side 2

SparseMatrixColorings.adjacency_graph — Function

adjacency_graph(H::AbstractMatrix)

Return a Graph representing the nonzeros of a symmetric matrix (typically a Hessian matrix).

The adjacency graph of a symmetrix matric A ∈ ℝ^{n × n} is G(A) = (V, E) where

V = 1:n is the set of rows or columns i/j

(i, j) ∈ E whenever A[i, j] ≠ 0 and i ≠ j

SparseMatrixColorings.bipartite_graph — Function

bipartite_graph(J::AbstractMatrix)

Return a BipartiteGraph representing the nonzeros of a non-symmetric matrix (typically a Jacobian matrix).

The bipartite graph of a matrix A ∈ ℝ^{m × n} is Gb(A) = (V₁, V₂, E) where

V₁ = 1:m is the set of rows i

V₂ = 1:n is the set of columns j

(i, j) ∈ E whenever A[i, j] ≠ 0

SparseMatrixColorings.neighbors — Function

neighbors(bg::BipartiteGraph, Val(side), v::Integer)

Return the neighbors of v (a vertex from the specified side, 1 or 2), in the graph bg.

SparseMatrixColorings.vertices — Function

vertices(bg::BipartiteGraph, Val(side))

Return the list of vertices of bg from the specified side as a range 1:n.

Coloring

SparseMatrixColorings.partial_distance2_coloring — Function

partial_distance2_coloring(bg::BipartiteGraph, ::Val{side}, order::AbstractOrder)

Compute a distance-2 coloring of the given side (1 or 2) in the bipartite graph bg and return a vector of integer colors.

A distance-2 coloring is such that two vertices have different colors if they are at distance at most 2.

The vertices are colored in a greedy fashion, following the order supplied.

See also

BipartiteGraph

AbstractOrder

SparseMatrixColorings.star_coloring1 — Function

star_coloring1(g::Graph, order::AbstractOrder)

Compute a star coloring of all vertices in the adjacency graph g and return a vector of integer colors.

A star coloring is a distance-1 coloring such that every path on 4 vertices uses at least 3 colors.

The vertices are colored in a greedy fashion, following the order supplied.

See also

Graph

AbstractOrder

Testing

SparseMatrixColorings.check_structurally_orthogonal_columns — Function

check_structurally_orthogonal_columns(
    A::AbstractMatrix, colors::AbstractVector{<:Integer}
    verbose=false
)

Return true if coloring the columns of the matrix A with the vector colors results in a partition that is structurally orthogonal, and false otherwise.

A partition of the columns of a matrix A is structurally orthogonal if, for every nonzero element A[i, j], the group containing column A[:, j] has no other column with a nonzero in row i.

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

SparseMatrixColorings.check_structurally_orthogonal_rows — Function

check_structurally_orthogonal_rows(
    A::AbstractMatrix, colors::AbstractVector{<:Integer};
    verbose=false
)

Return true if coloring the rows of the matrix A with the vector colors results in a partition that is structurally orthogonal, and false otherwise.

A partition of the rows of a matrix A is structurally orthogonal if, for every nonzero element A[i, j], the group containing row A[i, :] has no other row with a nonzero in column j.

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

SparseMatrixColorings.check_symmetrically_orthogonal — Function

check_symmetrically_orthogonal(
    A::AbstractMatrix, colors::AbstractVector{<:Integer};
    verbose=false
)

Return true if coloring the columns of the symmetric matrix A with the vector colors results in a partition that is symmetrically orthogonal, and false otherwise.

A partition of the columns of a symmetrix matrix A is symmetrically orthogonal if, for every nonzero element A[i, j], either

the group containing the column A[:, j] has no other column with a nonzero in row i

the group containing the column A[:, i] has no other column with a nonzero in row j

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.