API reference

Public

SparseMatrixColorings.SparseMatrixColorings — Module

SparseMatrixColorings

Coloring algorithms for sparse Jacobian and Hessian matrices.

The algorithms implemented in this package are all taken from the following survey:

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

Some parts of the survey (like definitions and theorems) are also copied verbatim or referred to by their number in the documentation.

SparseMatrixColorings.GreedyColoringAlgorithm — Type

GreedyColoringAlgorithm <: ADTypes.AbstractColoringAlgorithm

Matrix coloring algorithm for sparse Jacobians and Hessians.

Compatible with the ADTypes.jl coloring framework.

Implements

ADTypes.column_coloring with a partial distance-2 coloring of the bipartite graph

ADTypes.row_coloring with a partial distance-2 coloring of the bipartite graph

ADTypes.symmetric_coloring with a star coloring of the adjacency graph

Private

SparseMatrixColorings.AdjacencyGraph — Type

AdjacencyGraph

Represent a graph between the columns of a symmetric n × n matrix A with nonzero diagonal elements.

This graph is defined as G = (C, E) where C = 1:n is the set of columns and (i, j) ∈ E whenever A[i, j] is nonzero for some j ∈ 1:m, j ≠ i.

Fields

A_colmajor::AbstractMatrix: output of col_major applied to A

SparseMatrixColorings.BipartiteGraph — Type

BipartiteGraph

Represent a bipartite graph between the rows and the columns of a non-symmetric m × n matrix A.

This graph is defined as G = (R, C, E) where R = 1:m is the set of row indices, C = 1:n is the set of column indices and (i, j) ∈ E whenever A[i, j] is nonzero.

Fields

A_colmajor::AbstractMatrix: output of col_major applied to A (useful to get neighbors of a column)

A_rowmajor::AbstractMatrix: output of row_major applied to A (useful to get neighbors of a row)

SparseMatrixColorings.check_structurally_orthogonal_columns — Method

check_structurally_orthogonal_columns(A, colors; verbose=false)

Return true if coloring the columns of the matrix A with the vector colors results in a partition that is structurally orthogonal, and false otherwise.

Def 3.2: A partition of the columns of a matrix A is structurally orthogonal if, for every nonzero element A[i, j], the group containing column A[:, j] has no other column with a nonzero in row i.

Thm 3.5: The function distance2_column_coloring applied to the BipartiteGraph of A should return a suitable coloring.

SparseMatrixColorings.check_structurally_orthogonal_rows — Method

check_structurally_orthogonal_rows(A, colors; verbose=false)

Return true if coloring the rows of the matrix A with the vector colors results in a partition that is structurally orthogonal, and false otherwise.

Def 3.2: A partition of the rows of a matrix A is structurally orthogonal if, for every nonzero element A[i, j], the group containing row A[i, :] has no other row with a nonzero in column j.

Thm 3.5: The function distance2_row_coloring applied to the BipartiteGraph of A should return a suitable coloring.

SparseMatrixColorings.check_symmetrically_orthogonal — Method

check_symmetrically_orthogonal(A, colors; verbose=false)

Return true if coloring the columns of the symmetric matrix A with the vector colors results in a partition that is symmetrically orthogonal, and false otherwise.

Def 4.2: A partition of the columns of a symmetrix matrix A is symmetrically orthogonal if, for every nonzero element A[i, j], either

the group containing the column A[:, j] has no other column with a nonzero in row i

the group containing the column A[:, i] has no other column with a nonzero in row j

SparseMatrixColorings.col_major — Method

col_major(A::AbstractMatrix)

Construct a column-major representation of the matrix A.

SparseMatrixColorings.distance2_column_coloring — Method

distance2_column_coloring(g::BipartiteGraph)

Compute a distance-2 coloring of the column vertices in the BipartiteGraph g and return a vector of integer colors.

A distance-2 coloring is such that two vertices have different colors if they are at distance at most 2.

The algorithm used is the greedy Algorithm 3.2.

SparseMatrixColorings.distance2_row_coloring — Method

distance2_row_coloring(g::BipartiteGraph)

Compute a distance-2 coloring of the row vertices in the BipartiteGraph g and return a vector of integer colors.

A distance-2 coloring is such that two vertices have different colors if they are at distance at most 2.

The algorithm used is the greedy Algorithm 3.2.

SparseMatrixColorings.row_major — Method

row_major(A::AbstractMatrix)

Construct a row-major representation of the matrix A.

SparseMatrixColorings.star_coloring — Method

star_coloring(g::BipartiteGraph)

Compute a star coloring of the column vertices in the AdjacencyGraph g and return a vector of integer colors.

Def 4.5: A star coloring is a distance-1 coloring such that every path on four vertices uses at least three colors.

The algorithm used is the greedy Algorithm 4.1.