API reference

Public, exported

SparseMatrixColorings.SparseMatrixColorings — Module

SparseMatrixColorings

SparseMatrixColorings.jl

Coloring algorithms for sparse Jacobian and Hessian matrices.

Getting started

To install this package, run the following in a Julia Pkg REPL:

pkg> add SparseMatrixColorings

Background

The algorithms implemented in this package are taken from the following articles:

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

New Acyclic and Star Coloring Algorithms with Application to Computing Hessians, Gebremedhin et al. (2007)

Efficient Computation of Sparse Hessians Using Coloring and Automatic Differentiation, Gebremedhin et al. (2009)

ColPack: Software for graph coloring and related problems in scientific computing, Gebremedhin et al. (2013)

Some parts of the articles (like definitions) are thus copied verbatim in the documentation.

Alternatives

ColPack.jl: a Julia interface to the C++ library ColPack

SparseDiffTools.jl: contains Julia implementations of some coloring algorithms

Coloring algorithms

SparseMatrixColorings.GreedyColoringAlgorithm — Type

GreedyColoringAlgorithm <: ADTypes.AbstractColoringAlgorithm

Greedy coloring algorithm for sparse Jacobians and Hessians, with configurable vertex order.

Compatible with the ADTypes.jl coloring framework.

Constructor

GreedyColoringAlgorithm(order::AbstractOrder=NaturalOrder())

See also

AbstractOrder

SparseMatrixColorings.column_coloring_detailed — Function

column_coloring_detailed(S::AbstractMatrix, algo::GreedyColoringAlgorithm)

Compute a partial distance-2 coloring of the columns in the bipartite graph of the matrix S, return an AbstractColoringResult.

Example

using SparseMatrixColorings, SparseArrays

algo = GreedyColoringAlgorithm(SparseMatrixColorings.LargestFirst())

S = sparse([
    0 0 1 1 0
    1 0 0 0 1
    0 1 1 0 0
    0 1 1 0 1
])

column_colors(column_coloring_detailed(S, algo))

# output

5-element Vector{Int64}:
 1
 2
 1
 2
 3

SparseMatrixColorings.row_coloring_detailed — Function

row_coloring_detailed(S::AbstractMatrix, algo::GreedyColoringAlgorithm)

Compute a partial distance-2 coloring of the rows in the bipartite graph of the matrix S, return an AbstractColoringResult.

Example

using SparseMatrixColorings, SparseArrays

algo = GreedyColoringAlgorithm(SparseMatrixColorings.LargestFirst())

S = sparse([
    0 0 1 1 0
    1 0 0 0 1
    0 1 1 0 0
    0 1 1 0 1
])

row_colors(row_coloring_detailed(S, algo))

# output

4-element Vector{Int64}:
 2
 2
 3
 1

SparseMatrixColorings.symmetric_coloring_detailed — Function

symmetric_coloring(S::AbstractMatrix, algo::GreedyColoringAlgorithm)

Compute a star coloring of the columns in the adjacency graph of the symmetric matrix S, return an AbstractColoringResult.

Example

using SparseMatrixColorings, SparseArrays

algo = GreedyColoringAlgorithm(SparseMatrixColorings.LargestFirst())

S = sparse([
    1 1 1 1
    1 1 0 0
    1 0 1 0
    1 0 0 1
])

column_colors(symmetric_coloring_detailed(S, algo))

# output

4-element Vector{Int64}:
 1
 2
 2
 2

Result analysis

SparseMatrixColorings.AbstractColoringResult — Type

AbstractColoringResult{
    partition,
    symmetric,
    decompression,
    M<:AbstractMatrix
}

Abstract type for the detailed result of a coloring algorithm.

Type parameters

partition::Symbol: either :column, :row or :bidirectional

symmetric::Bool: either true or false

decompression::Symbol: either :direct or :substitution

M: type of the matrix that was colored

Applicable methods

column_colors and column_groups (for a :column or :bidirectional partition)

row_colors and row_groups (for a :row or :bidirectional partition)

get_matrix

SparseMatrixColorings.column_colors — Function

column_colors(result::AbstractColoringResult)

Return a vector color of integer colors, one for each column of the colored matrix.

SparseMatrixColorings.row_colors — Function

row_colors(result::AbstractColoringResult)

Return a vector color of integer colors, one for each row of the colored matrix.

SparseMatrixColorings.column_groups — Function

column_groups(result::AbstractColoringResult)

Return a vector group such that for every color c, group[c] contains the indices of all columns that are colored with c.

SparseMatrixColorings.row_groups — Function

row_groups(result::AbstractColoringResult)

Return a vector group such that for every color c, group[c] contains the indices of all rows that are colored with c.

Public, not exported

Decompression

SparseMatrixColorings.decompress — Function

decompress(B::AbstractMatrix, result::AbstractColoringResult)

Decompress B out-of-place into a new matrix A, given a coloring result of the sparsity pattern of A.

See also

AbstractColoringResult

SparseMatrixColorings.decompress! — Function

decompress!(
    A::AbstractMatrix, B::AbstractMatrix,
    result::AbstractColoringResult,
)

Decompress B in-place into an existing matrix A, given a coloring result of the sparsity pattern of A.

See also

AbstractColoringResult

Orders

SparseMatrixColorings.AbstractOrder — Type

AbstractOrder

Abstract supertype for vertex ordering schemes.

Subtypes

NaturalOrder

RandomOrder

LargestFirst

SparseMatrixColorings.NaturalOrder — Type

NaturalOrder()

Order vertices as they come in the graph.

SparseMatrixColorings.RandomOrder — Type

RandomOrder(rng=default_rng())

Order vertices with a random permutation.

SparseMatrixColorings.LargestFirst — Type

LargestFirst()

Order vertices by decreasing degree.

Private

Graph storage

SparseMatrixColorings.Graph — Type

Graph{T}

Undirected graph structure stored in Compressed Sparse Column (CSC) format.

Fields

colptr::Vector{T}: same as for SparseMatrixCSC

rowval::Vector{T}: same as for SparseMatrixCSC

SparseMatrixColorings.BipartiteGraph — Type

BipartiteGraph{T}

Undirected bipartite graph structure stored in bidirectional Compressed Sparse Column format (redundancy allows for faster access).

A bipartite graph has two "sides", which we number 1 and 2.

Fields

g1::Graph{T}: contains the neighbors for vertices on side 1

g2::Graph{T}: contains the neighbors for vertices on side 2

SparseMatrixColorings.vertices — Function

vertices(bg::BipartiteGraph, Val(side))

Return the list of vertices of bg from the specified side as a range 1:n.

SparseMatrixColorings.neighbors — Function

neighbors(bg::BipartiteGraph, Val(side), v::Integer)

Return the neighbors of v (a vertex from the specified side, 1 or 2), in the graph bg.

SparseMatrixColorings.adjacency_graph — Function

adjacency_graph(H::AbstractMatrix)

Return a Graph representing the nonzeros of a symmetric matrix (typically a Hessian matrix).

The adjacency graph of a symmetrix matric A ∈ ℝ^{n × n} is G(A) = (V, E) where

V = 1:n is the set of rows or columns i/j

(i, j) ∈ E whenever A[i, j] ≠ 0 and i ≠ j

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

SparseMatrixColorings.bipartite_graph — Function

bipartite_graph(J::AbstractMatrix)

Return a BipartiteGraph representing the nonzeros of a non-symmetric matrix (typically a Jacobian matrix).

The bipartite graph of a matrix A ∈ ℝ^{m × n} is Gb(A) = (V₁, V₂, E) where

V₁ = 1:m is the set of rows i

V₂ = 1:n is the set of columns j

(i, j) ∈ E whenever A[i, j] ≠ 0

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

Low-level coloring

SparseMatrixColorings.partial_distance2_coloring — Function

partial_distance2_coloring(bg::BipartiteGraph, ::Val{side}, order::AbstractOrder)

Compute a distance-2 coloring of the given side (1 or 2) in the bipartite graph bg and return a vector of integer colors.

A distance-2 coloring is such that two vertices have different colors if they are at distance at most 2.

The vertices are colored in a greedy fashion, following the order supplied.

See also

BipartiteGraph

AbstractOrder

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005), Algorithm 3.2

SparseMatrixColorings.symmetric_coefficient — Function

symmetric_coefficient(
    i::Integer, j::Integer,
    color::AbstractVector{<:Integer},
    group::AbstractVector{<:AbstractVector{<:Integer}},
    S::AbstractMatrix{Bool}
)

symmetric_coefficient(
    i::Integer, j::Integer,
    color::AbstractVector{<:Integer},
    star_set::StarSet
)

Return the indices (k, c) such that A[i, j] = B[k, c], where A is the uncompressed symmetric matrix and B is the column-compressed matrix.

The first version corresponds to algorithm DirectRecover1 in the paper, the second to DirectRecover2.

References

Efficient Computation of Sparse Hessians Using Coloring and Automatic Differentiation, Gebremedhin et al. (2009), Figures 2 and 3

SparseMatrixColorings.star_coloring — Function

star_coloring(g::Graph, order::AbstractOrder)

Compute a star coloring of all vertices in the adjacency graph g and return a tuple (color, star_set), where

color is the vector of integer colors

star_set is a StarSet encoding the set of 2-colored stars

A star coloring is a distance-1 coloring such that every path on 4 vertices uses at least 3 colors.

The vertices are colored in a greedy fashion, following the order supplied.

See also

Graph

AbstractOrder

References

New Acyclic and Star Coloring Algorithms with Application to Computing Hessians, Gebremedhin et al. (2007), Algorithm 4.1

SparseMatrixColorings.StarSet — Type

StarSet

Encode a set of 2-colored stars resulting from the star coloring algorithm.

Fields

The fields are not part of the public API, even though the type is.

star::Dict{Tuple{Int,Int},Int}: a mapping from edges (pair of vertices) their to star index

hub::Vector{Int}: a mapping from star indices to their hub (the hub is 0 if the star only contains one edge)

References

New Acyclic and Star Coloring Algorithms with Application to Computing Hessians, Gebremedhin et al. (2007), Algorithm 4.1

SparseMatrixColorings.group_by_color — Function

group_by_color(color::Vector{Int})

Create group::Vector{Vector{Int}} such that i ∈ group[c] iff color[i] == c.

Assumes the colors are contiguously numbered from 1 to some cmax.

SparseMatrixColorings.get_matrix — Function

get_matrix(result::AbstractColoringResult)

Return the matrix that was colored.

Testing

SparseMatrixColorings.same_sparsity_pattern — Function

same_sparsity_pattern(A::AbstractMatrix, B::AbstractMatrix)

Perform a partial equality check on the sparsity patterns of A and B:

if the return is true, they might have the same sparsity pattern but we're not sure

if the return is false, they definitely don't have the same sparsity pattern

SparseMatrixColorings.directly_recoverable_columns — Function

directly_recoverable_columns(
    A::AbstractMatrix, color::AbstractVector{<:Integer}
    verbose=false
)

Return true if coloring the columns of the symmetric matrix A with the vector color results in a column-compressed representation that preserves every unique value, thus making direct recovery possible.

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

SparseMatrixColorings.symmetrically_orthogonal_columns — Function

symmetrically_orthogonal_columns(
    A::AbstractMatrix, color::AbstractVector{<:Integer};
    verbose=false
)

Return true if coloring the columns of the symmetric matrix A with the vector color results in a partition that is symmetrically orthogonal, and false otherwise.

A partition of the columns of a symmetrix matrix A is symmetrically orthogonal if, for every nonzero element A[i, j], either of the following statements holds:

the group containing the column A[:, j] has no other column with a nonzero in row i

the group containing the column A[:, i] has no other column with a nonzero in row j

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

SparseMatrixColorings.structurally_orthogonal_columns — Function

structurally_orthogonal_columns(
    A::AbstractMatrix, color::AbstractVector{<:Integer}
    verbose=false
)

Return true if coloring the columns of the matrix A with the vector color results in a partition that is structurally orthogonal, and false otherwise.

A partition of the columns of a matrix A is structurally orthogonal if, for every nonzero element A[i, j], the group containing column A[:, j] has no other column with a nonzero in row i.

Warning

This function is not coded with efficiency in mind, it is designed for small-scale tests.

References

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

Matrix handling

SparseMatrixColorings.respectful_similar — Function

respectful_similar(A::AbstractMatrix)
respectful_similar(A::AbstractMatrix, ::Type{T})

Like Base.similar but returns a transpose or adjoint when A is a transpose or adjoint.

SparseMatrixColorings.matrix_versions — Function

matrix_versions(A::AbstractMatrix)

Return various versions of the same matrix:

dense and sparse

transpose and adjoint

Used for internal testing.